requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-import-export==2.7.1
et-xmlfile==1.1.0
Faker==11.1.0
ipython==7.30.1
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.3
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.0
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
traitlets==5.1.1
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
django-autocomplete-light
sorl-thumbnail
python-memcached
